+++ src/Services/Rating/Rating.Application/Services/RatingService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using Rating.Application.IRepositories;
 using Rating.Application.IServices;
@@ -14,25 +15,41 @@ namespace Rating.Application.Services
 		private readonly IUserRepository _userRepository;
 		private readonly IMapper _mapper;
 		private readonly IRedisService<List<Domain.Models.Rating?>> _redis;
+		private readonly ILogger<RatingService> _logger;
 
 		public RatingService(IRatingRepository ratingRepository,
 			IFilmRepository filmRepository, IUserRepository userRepository,
 			IMapper mapper, 
-			IRedisService<List<Domain.Models.Rating?>> redis)
+			IRedisService<List<Domain.Models.Rating?>> redis,
+			ILogger<RatingService> logger)
 		{
 			_ratingRepository = ratingRepository;
 			_filmRepository = filmRepository;
 			_userRepository = userRepository;
 			_mapper = mapper;
 			_redis = redis;
+			_logger = logger;
 		}
 
 		public async Task<Domain.Models.Rating> CreateRatingAsync(CreateRatingDTO ratingDTO)
 		{
-			var film = await _filmRepository.GetFilmByNameAsync(ratingDTO.FilmName)
-				?? throw new NotFoundException("Not Found this filmName!");
-			var user = await _userRepository.GetUserByLoginAsync(ratingDTO.UserLogin)
-				?? throw new NotFoundException("Not Found this user!");
+			var film = await _filmRepository.GetFilmByNameAsync(ratingDTO.FilmName);
+
+			if (film is null)
+			{
+				_logger.LogError("Not Found this filmName!");
+
+				throw new NotFoundException("Not Found this filmName!");
+			}
+
+			var user = await _userRepository.GetUserByLoginAsync(ratingDTO.UserLogin);
+
+			if (user is null)
+			{
+				_logger.LogError("Not Found this user!");
+
+				throw new NotFoundException("Not Found this user!");
+			}
 
 			var rating = new Domain.Models.Rating()
 			{
@@ -44,16 +61,26 @@ namespace Rating.Application.Services
 			await _ratingRepository.CreateAsync(rating);
 			await _ratingRepository.SaveAsync();
 
+			_logger.LogInformation($"Rating was successfully created. Id: {rating.Id}");
+
 			return rating;
 		}
 
 		public async Task DeleteRating(int id)
 		{
-			var rating = await _ratingRepository.GetByIdAsync(id)
-				?? throw new NotFoundException("Can not found this rating!");
+			var rating = await _ratingRepository.GetByIdAsync(id);
+
+			if (rating is null)
+			{
+				_logger.LogError($"Can not found this rating! Id: {id}");
+
+				throw new NotFoundException("Can not found this rating!");
+			}
 
 			_ratingRepository.Delete(rating);
 			await _ratingRepository.SaveAsync();
+
+			_logger.LogInformation($"Rating was successfully deleted. Id: {id}");
 		}
 
 		public async Task<List<RatingDTO>> GetAllRatingsAsync()
@@ -67,12 +94,16 @@ namespace Rating.Application.Services
 
 				await _redis.SetAsync(ratings);
 
+				_logger.LogError("All ratings was successfully received and added to Redis.");
+
 				return ratingDtos;
 			}
 			else
 			{
 				var ratings = JsonConvert.DeserializeObject<List<Domain.Models.Rating?>>(serializedRatingList);
 				var ratingDtos = _mapper.Map<List<RatingDTO>>(ratings);
+
+				_logger.LogError("All ratings was successfully received.");
 
 				return ratingDtos;
 			}
@@ -83,18 +114,28 @@ namespace Rating.Application.Services
 			int ratingSum = await _ratingRepository.GetSumRatingForFilmNameAsync(filmName);
 			int count = await _ratingRepository.GetCountOfRatedUsers(filmName);
 
+			_logger.LogError($"Rating by filmName was successfully received.");
+
 			return ratingSum / count;
 		}
 
 		public async Task UpdateRating(RatingDTO ratingDTO)
 		{
-			var rating = await _ratingRepository.GetByIdAsync(ratingDTO.Id)
-				?? throw new NotFoundException("Can not found this rating!");
+			var rating = await _ratingRepository.GetByIdAsync(ratingDTO.Id);
+
+			if (rating is null)
+			{
+				_logger.LogError("Can not found this rating!");
+
+				throw new NotFoundException("Can not found this rating!");
+			}
 
 			rating = _mapper.Map<Domain.Models.Rating>(ratingDTO);
 
 			_ratingRepository.Update(rating);
 			await _ratingRepository.SaveAsync();
+
+			_logger.LogInformation($"Rating was successfully updated. Id: {rating.Id}");
 		}
 	}
 }
